+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Next Up
 
+- Backgrounds now capture equipment during character option munch!
+
 # 7.3.9
 
 - More improvements to background equipment and feat parsing. @crimsonedge